+++ src/org/helioviewer/jhv/gui/components/MaterialDesign.java
@@ -31,6 +31,7 @@ enum MaterialDesign {
     PLUS_CIRCLE("mdi-plus-circle", '\uf417'),
     PLUS_ONE("mdi-plus-one", '\uf41b'),
     RECORD("mdi-record", '\uf44a'),
+    REFRESH("mdi-refresh", '\uf450'),
     ROTATE_3D("mdi-rotate-3d", '\uf464'),
     ROTATE_90("mdi-rotate-90", '\uf6a9'),
     SHAPE_POLYGON_PLUS("mdi-shape-polygon-plus", '\uf65e'),